add util fn for renaming
