+++ examples/pytorch/gcn/README.md
@@ -16,16 +16,17 @@ pip install torch requests
 
 Codes
 -----
-The folder contains two implementations of GCN. `gcn.py` uses user-defined
-message and reduce functions. `gcn_spmv.py` uses DGL's builtin functions so
-SPMV optimization could be applied.
+The folder contains three implementations of GCN:
+- `gcn.py` uses DGL's predefined graph convolution module.
+- `gcn_mp.py` uses user-defined message and reduce functions.
+- `gcn_spmv.py` uses DGL's builtin functions so SPMV optimization could be applied.
 
 Results
 -------
 
 Run with following (available dataset: "cora", "citeseer", "pubmed")
 ```bash
-python gcn_spmv.py --dataset cora --gpu 0
+python gcn.py --dataset cora --gpu 0
 ```
 
 * cora: ~0.810 (0.79-0.83) (paper: 0.815)
