--- tests/pytorch/placeholder.py
@@ -1,3 +1,0 @@
-"""
-Placeholder file for framework-specific test
-"""
